Fix flakiness of client_failover-itest
The reason this test was flaky is that there is a race between
determining that all of the replicas in a config have replicated an opid
and that opid being committed.

This patch exposes getting the committed OpId from Consensus as an RPC
(previously, only the last replicated opid was exposed). It also adds a
helper function for testing purposes to wait until a specific OpId has
been committed.

Change-Id: I56ea7523e212c1b0f67b1498ba5139e4c5537519