Add a design doc for rpc retry/failover semantics
This adds a first version of an rpc retry failover doc for folks
to review. Note that while it "chooses" a path (overly using hypotheticals
makes for long, nebulous, docs) that path is up for discussion so feel free
to weigh in if you feel that something should be different/implemented in
a different order.

Change-Id: Idc2aa40486153b39724e1c9bd09c626b829274c6