KUDU-1377 (part 4): Ignore corrupted trailing records on LBM metadata
This causes the LogBlockManager to tolerate trailing truncated records
in container metadata files. This can occur in a crash or full-disk
scenario. The way it works is that if the metadata file is detected to
have only a partial trailing record, which is assumed to have been an
in-progress IO at the time of a crash and is truncated. We continue
writing to the file at that point.

Adds a test that corrupts a LBC metadata file in various ways.

Change-Id: I0b88ba051b390b6a2587eecdd205c478f1edccb4