Add Google Breakpad support to Kudu
Breakpad part of the crash-reporting infrastructure open-sourced by
Google. It is used by Mozilla Firefox and Google Chrome.

When a process crashes, Breakpad writes a small dump file called a
minidump to disk. Since these files are typically only a few MB in size,
this allows users to share critical crash information with developers
for offline analysis.

Changes in this patch:

* Add breakpad to thirdparty.
* Add breakpad source archive creation script (breakpad does not do releases).
* Add "-system $PREFIX/includes/breakpad" to the build command line,
  because breakpad does not properly namespace its includes (however it
  installs them to their own directory).
* Pull in (modified) util/minidump.{cc,h} from Apache Impala (incubating).
  The modifications mostly consist of removing use of boost::filesystem
  and conformance to the Kudu code style guidelines.
* Enable breakpad support for the kudu-tserver and kudu-master process
  startup routines.

Change-Id: I495695cc38b75377f20b0497093a81ed5baa887f