log: Mark allocation finished even if allocation had an error
The problem this patch fixes is that if a disk preallocation fails we
will enter a "stuck" state where we cannot preallocate a new segment.
Errors to allocate or append should be fatal.

A test exercising this code path is in a follow-up patch related to disk
reservations. Since this is a bug fix, it seemed cleaner to separate it
out into its own commit.

Change-Id: If22bf946a42d0ec32c35164acd9e6e6cef18dcc3