Remove the LocalConsensus implementation
This implementation was intended as scaffolding while we built a true
consensus implementation. Now that we have a full implementation of Raft
consensus, it is no longer needed.

Also update ConsensusMetadata validation and remove the use of the
RaftConfigPB.local field (marked as obsolete).

Additionally, convert alter_table-test to use RaftConsensus:

1. Wait for tablet to be elected leader before proceeding.
2. All of the timestamps must be incremented due to the initial no-op
   written to the WAL by the Raft leader.

Finally, update the following tests to specify themselves as voters when
creating a single-node config:

* consensus_meta-test
* raft_consensus_state-test
* tablet_bootstrap-test

Change-Id: I24e671dc732a1cdf4c453dfec61cefa0c573252c