compaction: Flush tablet metadata before updating stores
We should always flush the rowset metadata before making changes visible
to the delta tracker. In this case, because we are just doing a
compaction, it is likely not a crash consistency issue, however we
should always follow the same order of operations when operating on
tablet metadata.

Also, add LogPrefix() for the log lines printed from the delta tracker.

Change-Id: I958f4744f1b95c6a63e8c9105bd1c7552e43e6b9