RFC: subprocess: Don't require process-wide SIG_IGN for SIGPIPE
This patch is marked as RFC because the disadvantages of this
implementation may outweigh its advantages.

This patch implements an alternative approach to handling SIGPIPE while
forking a subprocess that doesn't require a process-wide signal
disposition of SIG_IGN for SIGPIPE.

One downside to this approach is that while it is less invasive than
ignoring SIGPIPE process-wide, there are opportunities to ignore
process-wide PIPE signals due to a particular race.

The way this patch works is that the thread that may generate a SIGPIPE
blocks the PIPE signal in its per-thread signal mask, performs the
potentially signal-inducing write() call on a pipe, consumes a
potentially pending SIGPIPE notification, and unblocks the delivery of
SIGPIPE.

This per-thread behavior is possible because SIGPIPE is delivered to the
thread that generated it. From the POSIX.1 spec section 2.10.14 Signals:

  The SIGPIPE signal shall be sent to a thread that attempts to send
  data on a socket that is no longer able to send (one that is no longer
  connected), except that the signal is suppressed if the MSG_NOSIGNAL
  flag is used in calls to send(), sendto(), and sendmsg(). Regardless
  of whether the generation of the signal is suppressed, the send
  operation shall fail with the [EPIPE] error.

One problem with the above approach, however, is that in between the
time that a SIGPIPE becomes pending on the thread blocking it and the
time that thread consumes that pending signal, delivery of SIGPIPE is
blocked process-wide because the signal is already pending.

Using this approach in general has the benefit that clients do not have
to set the disposition of SIGPIPE to SIG_IGN process-wide for Kudu's
sake, but it has the major downside that delivery of SIGPIPE becomes
unreliable and non-deterministic in cases where delivery of that signal
may be desired.

Change-Id: I4d10207e1bc09493bbf9131d4e622ace895d2186