Fix flaky test TestRecoverFromOpIdOverflow
This test is flaky because we race against the COMMIT message for the
first NO_OP in the WAL being written. It is currently hard to know when
the actual COMMIT message is written to the WAL so we use a workaround
to delete the first log segment before restarting the EMC in this test.

Change-Id: Ib382819307da04bb76d68d2c015dc0edd9f60267