WIP: Create parallelized loader Spark job
This patch adds a new tool called LoadRandomData that will load random
data into a table in a parallelized manner. Perhaps the tool should be
named something else so it would be straightforward to add a sequential
inserter that supports a range-partitioned table.

Marked WIP because it could use more testing.

Change-Id: I4434c9f5d709154037386b4c7be94045df162267