KUDU-2645: tablet: Add ghost row de-duplication support to MergeIterator
This patch adds support to the MergeIterator for de-duplicating ghost
rows found when a row with a particular primary key is deleted and
reinserted into a different rowset. These duplicates may be returned
when the 'include_deleted_rows' option is enabled in the row iterator
options.

The result is that the UnionIterator will support returning deleted
rows, but will not de-duplicate them. In contrast, the MergeIterator
will always de-duplicate deleted rows if they are included in the result
set.

Change-Id: I00614b3fa5c6b4e7b620bb78489e24c5ad44daee