KUDU-2645: schema: Validate that defaults are provided for virtual columns
Virtual columns must not be nullable and must have read defaults.
Instead of crashing when we encounter violations of these requirements,
return a bad status code when mapping such a projection onto a tablet
schema.

TODO(KUDU-2692): Consider removing the requirement to specify a read
default on a virtual column by defining a built-in read default for each
virtual column type, such as a default of false for IS_DELETED.

Change-Id: I9144390f9ed02789a1ea161277b5f1567cfc2ffe