[consensus] Implement routing table
This patch implements a "routing table", which is actually more like a
routing tree, since there is a single-parent relationship among nodes
represented in the routing table, thus there is only one possible route
to each node at any given instance of the configuration. See the comment
at the top of routing.h for details on the assumptions made and the
behavior of the routing table.

Included is a simple unit test of the next-hop routing behavior.

Change-Id: I44e9ccfe4ca97ac24c30b7163778840e1d4b823e