[consensus][proxy] implement routing table
This patch implements a "routing table", which is actually more of a
routing tree, since there is a single-parent relationship among nodes
represented in the routing table, thus there is only one possible route
to each node at any given instance of the configuration. See the comment
at the top of routing.h for details on the assumptions made and the
behavior of the routing table.

Included is a simple unit test of the next-hop routing behavior.

This patch also implements a DurableRoutingTable which persists the
relevant routing metadata to disk. The DurableRoutingTable could use
unit tests, but for now has integration test coverage in a later patch
in this series.

The specification of the proxy routing topology is separated from
RaftConfigPB so that it can be managed separately from Raft config
change.

Some TODOs:

* Could use some lower level test coverage for DRT including a
  persistence test.
* Degrade to direct routing when proxy_from_uuid does not point to a
  valid member of the config. Right now it throws an error.

Change-Id: I44e9ccfe4ca97ac24c30b7163778840e1d4b823e